moved member attributes funcs
